+++ apps/dashboard/src/app/providers.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { getPublicEnvVar } from '@/lib/env';
+import { getPublicEnvVar } from '@stackframe/stack-shared/dist/utils/env';
 import { useStackApp, useUser } from '@stackframe/stack';
 import posthog from 'posthog-js';
 import { PostHogProvider } from 'posthog-js/react';
@@ -9,6 +9,12 @@ if (typeof window !== 'undefined') {
   const postHogKey = getPublicEnvVar('NEXT_PUBLIC_POSTHOG_KEY') ?? "phc_vIUFi0HzHo7oV26OsaZbUASqxvs8qOmap1UBYAutU4k";
   if (postHogKey.length > 5) {
     posthog.init(postHogKey, {
+      session_recording: {
+        maskAllInputs: false,
+        maskInputOptions: {
+          password: true,
+        },
+      },
       api_host: "/consume",
       ui_host: "https://eu.i.posthog.com",
       capture_pageview: false,