+++ examples/js-example/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stackframe/js-example
 
+## 2.7.14
+
+### Patch Changes
+
+- @stackframe/js@2.7.14
+
 ## 2.7.13
 
 ### Patch Changes